backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
openai==1.3.0
pyairtable==2.1.0
httpx==0.25.1
pydantic-settings==2.1.0

# Optional: Multi-agent frameworks (uncomment if needed)
# langgraph==0.0.20
# crewai==0.1.0
